django/0909/requirements.txt
argon2-cffi==20.1.0
asgiref==3.4.1
async-generator==1.10
attrs==21.2.0
backcall==0.2.0
beautifulsoup4==4.9.3
bleach==3.3.1
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.3
colorama==0.4.4
debugpy==1.4.1
decorator==5.0.9
defusedxml==0.7.1
Django==3.2.6
django-appconf==1.0.4
django-bootstrap-v5==1.0.5
django-cleanup==5.2.0
django-extensions==3.1.3
django-imagekit==4.0.2
entrypoints==0.3
idna==3.2
ipykernel==6.0.3
ipython==7.25.0
ipython-genutils==0.2.0
jedi==0.18.0
Jinja2==3.0.1
jsonschema==3.2.0
jupyter-client==6.1.12
jupyter-contrib-core==0.3.3
jupyter-contrib-nbextensions==0.5.1
jupyter-core==4.7.1
jupyter-highlight-selected-word==0.2.0
jupyter-latex-envs==1.4.6
jupyter-nbextensions-configurator==0.4.1
jupyterlab-pygments==0.1.2
lxml==4.6.3
MarkupSafe==2.0.1
matplotlib-inline==0.1.2
mistune==0.8.4
nbclient==0.5.3
nbconvert==6.1.0
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.4.0
packaging==21.0
pandocfilters==1.4.3
parso==0.8.2
pickleshare==0.7.5
pilkit==2.0
Pillow==8.3.2
prometheus-client==0.11.0
prompt-toolkit==3.0.19
pycparser==2.20
Pygments==2.9.0
pyparsing==2.4.7
pyrsistent==0.18.0
python-dateutil==2.8.2
pytz==2021.1
pywin32==301
pywinpty==1.1.3
PyYAML==5.4.1
pyzmq==22.1.0
requests==2.26.0
Send2Trash==1.7.1
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.1
terminado==0.10.1
testpath==0.5.0
tornado==6.1
traitlets==5.0.5
urllib3==1.26.6
wcwidth==0.2.5
webencodings==0.5.1
secret.json